Move binary propagator and base64 format to contrib.
* Moved them as they are not part of the spec for now and should not be in main crate when GA.

* Also restructure the contrib into two submodule, one for tracing and one for metric.
