+++ opentelemetry-contrib/Cargo.toml
@@ -20,6 +20,8 @@ rustdoc-args = ["--cfg", "docsrs"]
 
 [features]
 default = []
+base64_format = ["base64", "binary_propagator"]
+binary_propagator = []
 datadog = ["indexmap", "rmp", "async-trait"]
 reqwest-blocking-client = ["reqwest/blocking", "opentelemetry/reqwest"]
 reqwest-client = ["reqwest", "opentelemetry/reqwest"]
@@ -34,6 +36,7 @@ lazy_static = "1.4"
 reqwest = { version = "0.10", optional = true }
 surf = { version = "2.0", optional = true }
 http = "0.2"
+base64 = { version = "0.13", optional = true }
 
 [dev-dependencies]
 base64 = "0.13"
