+++ opentelemetry-contrib/src/lib.rs
@@ -10,6 +10,8 @@
 //! The following crate feature flags are available:
 //!
 //! * `datadog`: Adds a Datadog trace exporter.
+//! * `binary-propagator`: experimental binary propagator to propagate trace context using binary format.
+//! * `base64-format`: enables base64 format support for binary propagators.
 //! * `reqwest-blocking-client`: Export spans using the reqwest blocking http
 //!   client.
 //! * `reqwest-client`: Export spans using the reqwest non-blocking http client.
@@ -29,6 +31,4 @@
 )]
 #![cfg_attr(test, deny(warnings))]
 
-#[cfg(feature = "datadog")]
-#[cfg_attr(docsrs, doc(cfg(feature = "datadog")))]
-pub mod datadog;
+pub mod trace;
