+++ opentelemetry-contrib/src/trace/exporter/datadog/model/v03.rs
@@ -1,4 +1,4 @@
-use crate::datadog::model::Error;
+use crate::trace::exporter::datadog::model::Error;
 use opentelemetry::exporter::trace;
 use opentelemetry::{Key, Value};
 use std::time::SystemTime;
