+++ opentelemetry-contrib/src/trace/mod.rs
@@ -1,0 +1,5 @@
+//! # Opentelemetry trace contrib
+//!
+
+pub mod exporter;
+pub mod propagator;
