+++ opentelemetry-contrib/src/trace/propagator/mod.rs
@@ -1,0 +1,9 @@
+//! # Opentelemetry propagator contrib
+//!
+//! This module provides propagators for third party vendor format or experimental propagators that
+//! aren't part of Opentelemetry.
+//!
+//! Currently, the following propagators are supported:
+//!
+//! * `binary_propagator`, propagating trace context in the binary format.
+pub mod binary;
