+++ opentelemetry/Cargo.toml
@@ -23,7 +23,6 @@ rustdoc-args = ["--cfg", "docsrs"]
 [dependencies]
 async-std = { version = "1.6", features = ["unstable"], default-features = false, optional = true  }
 async-trait = { version = "0.1", optional = true }
-base64 = { version = "0.13", optional = true }
 bincode = { version = "1.2", optional = true }
 dashmap = { version = "4.0.0-rc6", optional = true }
 fnv = { version = "1.0", optional = true }
@@ -48,11 +47,9 @@ tokio = { version = "0.2", features = ["full"] }
 
 [features]
 default = ["trace"]
-base64_format = ["base64", "binary_propagator"]
 trace = ["rand", "pin-project", "async-trait", "regex", "percent-encoding"]
 metrics = ["thiserror", "dashmap", "fnv"]
 serialize = ["serde", "bincode"]
-binary_propagator = []
 
 [[bench]]
 name = "trace"
