--- opentelemetry/src/experimental/mod.rs
@@ -1,3 +1,0 @@
-//! # OpenTelemetry Experimental Features
-
-pub mod propagation;
